+++ src/i18n.js
@@ -4,12 +4,14 @@ export const i18nValidator = {
     query: 'Query validator\'s balance...',
     accountBalance: 'Account Balance',
     validatorBalance: '(Eth2) Validator Balance',
+    xdaiBalance: 'xDai Chain Balance',
   },
   'zh_TW': {
     needAddr: '請傳入索引/地址，或是預先定義 SAIHUBOT_VALIDATOR 參數',
     query: '查詢驗證器餘額中...',
     accountBalance: '地址餘額',
     validatorBalance: '(Eth2) 驗證者餘額',
+    xdaiBalance: 'xDai 鏈上餘額',
   },
   props: [],
 };