+++ src/saihubot-cli-skill-chain.js
@@ -6,24 +6,11 @@ import AsciiBar from 'ascii-bar';
 import humanizeDuration from 'humanize-duration';
 import commaNumber from 'comma-number';
 import { t } from 'saihubot-cli-adapter/dist/i18n';
-import {getNodeURL} from './utils';
+import {ethFetch, xdaiFetch} from './utils';
 
 const ADDR = {
   ETH2_DEPOSIT: '0x00000000219ab540356cbb839cbe05303d7705fa',
 }
-
-const baseFetchOptions = {
-  method: "POST",
-  headers: {
-    "Content-Type": "application/json"
-  },
-};
-
-const ethFetch = (fetch, rpc) =>
-  fetch(getNodeURL(), {
-    ...baseFetchOptions,
-    body: rpc,
-  }).then(response => response.json());
 
 let idx = 1;
 
@@ -35,14 +22,14 @@ const rpcLastBlock = JSON.stringify({
   params: [],
 });
 
-const rpcEthBalance = (address) => JSON.stringify({
+export const rpcEthBalance = (address) => JSON.stringify({
   jsonrpc: '2.0',
   id: idx++,
   method: 'eth_getBalance',
   params: [`${address}`, "latest"],
 });
 
-/*const rpcTokenBalance = (address, token) => JSON.stringify({
+export const rpcTokenBalance = (address, token) => JSON.stringify({
   jsonrpc: '2.0',
   id: idx++,
   method: 'eth_call',
@@ -52,7 +39,7 @@ const rpcEthBalance = (address) => JSON.stringify({
     },
     'latest'
   ],
-});*/
+});
 
 const rpcGasPrice = () => JSON.stringify({
   jsonrpc: '2.0',
@@ -86,11 +73,44 @@ export const skillLastBlock = {
     robot.send(t('fetching', {i18n: this.i18n}));
     robot.render();
     ethFetch(robot.addons.fetch, rpcLastBlock)
-    .then(json => {
-      const msg = t('summary', {i18n: this.i18n, blocknum: parseInt(json.result)});
-      robot.send(msg);
-      robot.render();
-    })
+      .then(json => {
+        const msg = t('summary', {i18n: this.i18n, blocknum: parseInt(json.result)});
+        robot.send(msg);
+        robot.render();
+      });
+  },
+}
+
+/**
+ * Get the latest block number.
+ */
+export const skillLastXdaiBlock = {
+  name: 'lastblock-xdai',
+  help: 'lastblock-xdai|lastblockxdai|block-xdai|blockxdai - get the latest xDai block number',
+  requirements: {
+    addons: ['fetch'],
+  },
+  i18n: {
+    'en': {
+      fetching: 'Fetching data...',
+      summary: 'The latest xDai block is **{{blocknum}}**',
+    },
+    'zh_TW': {
+      fetching: '取得資料中...',
+      summary: '最新的xDai區塊是 **{{blocknum}}**',
+    },
+    props: ['blocknum']
+  },
+  rule: /^(last)?block-?xdai$/i,
+  action: function(robot, msg) {
+    robot.send(t('fetching', {i18n: this.i18n}));
+    robot.render();
+    xdaiFetch(robot.addons.fetch, rpcLastBlock)
+      .then(json => {
+        const msg = t('summary', {i18n: this.i18n, blocknum: parseInt(json.result)});
+        robot.send(msg);
+        robot.render();
+      });
   },
 }
 
@@ -304,5 +324,6 @@ export const skillGasFee = {
 }
 
 export const skillsETH2 = [skillEth2Stats, skillLastBlock, skillBeaconLastBlock];
-const skills = [...skillsETH2, skillGasFee];
+export const skillsXDAI = [skillLastXdaiBlock];
+const skills = [...skillsETH2, ...skillsXDAI, skillGasFee];
 export {skills};