+++ src/utils.js
@@ -25,6 +25,12 @@ const ETH_NODES = [
   'https://mainnet.eth.cloud.ava.do/', // AVADO
 ];
 
+// https://www.xdaichain.com/for-developers/developer-resources#json-rpc-endpoints
+const XDAI_NODES = [
+  'https://rpc.xdaichain.com/',
+  'https://xdai.poanetwork.dev/',
+];
+
 let cachedNodeURL = '';
 
 /**
@@ -37,3 +43,35 @@ export const getNodeURL = () => {
   cachedNodeURL = getConfig('NODE_URL', getRandomItem(ETH_NODES));
   return cachedNodeURL;
 }
+
+let cachedXdaiNodeUrl = '';
+
+/**
+ * Random pick a xdai node.
+ *
+ * can set yours via set SAIHUBOT_XDAI_NODE_URL environment variable.
+ */
+export const getXdaiNodeURL = () => {
+  if (cachedXdaiNodeUrl) return cachedXdaiNodeUrl;
+  cachedXdaiNodeUrl = getConfig('XDAI_NODE_URL', getRandomItem(XDAI_NODES));
+  return cachedXdaiNodeUrl;
+}
+
+const baseFetchOptions = {
+  method: "POST",
+  headers: {
+    "Content-Type": "application/json"
+  },
+};
+
+export const ethFetch = (fetch, body) =>
+  fetch(getNodeURL(), {
+    ...baseFetchOptions,
+    body,
+  }).then(response => response.json());
+
+export const xdaiFetch = (fetch, body) =>
+  fetch(getXdaiNodeURL(), {
+    ...baseFetchOptions,
+    body,
+  }).then(response => response.json());